Fixed memory overrun issues and incorrect semaphore use in httpd
