+++ src/hap_internal.h
@@ -28,7 +28,6 @@ struct hap_accessory {
     char pincode[HAP_PINCODE_LENGTH+1];
     char* name;
     char* vendor;
-    int port;
 
     enum hap_accessory_category category;
     uint32_t config_number;
